+++ src/test/java/store/model/promotion/PeriodTest.java
@@ -1,4 +1,4 @@
-package store.model;
+package store.model.promotion;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import store.model.promotion.Period;
 
 @DisplayName("== Period 테스트 ==")
 public class PeriodTest {
